GPX-580: restructure to flat
