--- pkg/utils/db.go
@@ -1,34 +1,0 @@
-package utils
-
-import (
-	"database/sql"
-	"fmt"
-	"github.com/go-sql-driver/mysql"
-	"os"
-)
-
-var DB *sql.DB
-
-func InitDB() {
-	if C.Db.Enabled {
-		password, err := os.ReadFile(C.Db.PasswordPath)
-		LogIfPanic("could not read db password", err)
-		cfg := mysql.Config{
-			User:                 C.Db.User,
-			Passwd:               string(password),
-			Net:                  "tcp",
-			AllowNativePasswords: true,
-			Addr:                 C.Db.Host + ":" + fmt.Sprint(C.Db.Port),
-			DBName:               C.Db.DbName,
-		}
-		// Get a database handle.
-		DB, err = sql.Open("mysql", cfg.FormatDSN())
-		LogIfPanic("Error opening database", err)
-	}
-}
-
-func CloseDB() {
-	if DB != nil {
-		LogIfPanic("Error closing result", DB.Close())
-	}
-}
